tinystdio: Fix x86/m68k 80-bit %a overflow when rounding
Formats without an implicit '1' for the MSB and which fill all four
bits of the top digit might actually overflow when rounding. Check for
that and shift right.  We can do this without loss of accuracy because
we just carried all the way from 'half' to the top digit of the value

Signed-off-by: Keith Packard <keithp@keithp.com>