meson: Add missing -Wl, prefix to cc.has_link_argument() call
This fixes the detection of ld.lld 15 which previously failed
unconditionally since the argument was being passed to the compiler and
not forwarded to the linker.
However, even with -Wl, this check fails since we are linking with
-nostartfiles and thus ld.lld complains about missing _start. We can work
around this problem by also passing --version, which ensures that we get an
error for unknown arguments and a zero exit code if all arguments exist.

Fixes https://github.com/picolibc/picolibc/issues/773