tinystdio: handling `"nan(n-sequence char)"` by `scanf()` family & `strtod()`
Quoting from C standard part 7.19.6.2

> `a,e,f,g` Matches an optionally signed floating-point number, infinity, or NaN, whose
format is the same as expected for the subject sequence of the `strtod`
function. The corresponding argument shall be a pointer to floating.

Regarding `strtod`, quoting from C standard part 7.20.1.3

> A character sequence `NAN` or `NAN(n-char-sequence)`, is interpreted as a quiet
NaN, if supported in the return type, else like a subject sequence part that does not have
the expected form; the meaning of the `n-char sequences` is implementation-defined.

Signed-off-by: Abdallah Abdelhafeez (Abdallahs70)