arm: Initialize all of the shadow SPs for non-M targets
Non-M processors have a wealth of stack pointers, one for each
interrupt entry point. As the only thing we're using the other entry
points for in the sample code is to catch errors, we'll just smash
the same stack for every one.

This allows us to stop setting the SP in the vectors themselves which
trashed one of the saved register values.

This is mixed in with code which uses ARM mode exception vectors on
'A' profile targets; attempting to set the TE bit on those doesn't
seem to work on QEMU.

Signed-off-by: Keith Packard <keithp@keithp.com>