Restore original ifdef in limits.h for the include_next of system's limits.h
Commit 2f5f8439bd61e93e0f42d8009324d8c5227153c1 made changes to the
preprocessor guard that gated the include_next directive for limits.h

Currently the guard looks something like this:

```
ifdef __GNUC__
 include_next<limits.h>
endif
```

The original guard, on the other hand, also checked for _GCC_LIMITS_H_.
This extra condition is necessary to build with Clang. This error
message is shown:

> fatal error: 'limits.h' file not found

Because include_next is run when it shouldn't.