Ensure invalid locale strings raise an exception
Updated `newlocale` function to handle empty, "C" and "POSIX" locale strings
when `_MB_CAPABLE` is not defined.
The function now returns the "C" locale structure for these valid locales.
For any other locale string, it sets `errno` to `EINVAL` and returns `NULL`,
ensuring that an exception is raised in `C++` when using `std::locale`.

Signed-off-by: Mostafa Salman <mostafas@synopsys.com>