libm: Improve powf accuracy
One of the constants, 'cp', was not correctly converted to extended
precision (using two floats) leading to reduced precision in the
computation of log2(x).

The new values use a larger value for cp_h which is closer to the
actual value.  That leaves a smaller magnitude for cp_l resulting in
increased precision for the result. These values were computed
according to the Veltkamp/Dekker method.

2/(3*ln2): 0x1.ec709dc3a03fd748p-1
Old cp_h:  0x1.ec7p-1
New cp_h:  0x1.ec8p-1

The desired value is much closer to the higher value than the lower.

Signed-off-by: Keith Packard <keithp@keithp.com>