picocrt: Adapt x86 crt0 code for old clang assembler
Before clang-12, the built-in x86 assembler seems to treat prefixes as
separate instructions. This meant that it wouldn't allow them on the
same source line as the rest of the instruction. Split out the 'cs'
prefix used in the x86 crt0 code to a separate line so that old clang
versions can also assemble the code.

Signed-off-by: Keith Packard <keithp@keithp.com>