Remove GCC libdir from startfile specs fragment
When GCC is processing the *startfile spec, every time it
builds a complete value, it searches for that in the
internal list of library directories and prepends the
directory name if it finds it. That happens even when
processing parameters to the 'getenv function.

While processing this:

%:getenv(GCC_EXEC_PREFIX ../../picolibc/arm-zephyr-eabi/lib/%*/%M)

if --picolibc-buildtype is 'release' and the multilib is 'thumb/v7-m/nofp',
you get these two arguments:

	GCC_EXEC_PREFIX
	../../picolibc/arm-zephyr-eabi/lib/release/thumb/v7-m/nofp

Because *startfile generates a filename, gcc will take that second argument
and perform a file search, prepending the GCC lib directory to it so you get

	GCC_EXEC_PREFIX
	/opt/arm-zephyr-eabi/bin/../lib/gcc/../../picolibc/arm-zephyr-eabi/lib/release/thumb/v7-m/nofp

Now we evaluate the getenv function itself and end up prepending the
lib directory all over again. What that means is we can stop using
getenv or %R in the *startfile spec and let gcc do the search and
stick the lib dir on for us.

To do this, I added another formatter, specs_rel_format, which
generates paths relative to the lib dir while the existing one still
generates absolute paths. I renamed the absolute one specs_abs_format.

Signed-off-by: Keith Packard <keithp@keithp.com>