string: Add memory barrier to memset_s
The memset_s spec talks about how it's different from memset and has
stronger guarantees. I can't see how that actually requires the writes
if the memory is released afterwards (either through a call to free or
a function return), but it seems like the spec might encourage people
to think that it does. Adding a memory barrier ensures the compiler
will not elide the actual memory operations even when built with LTO.

Signed-off-by: Keith Packard <keithp@keithp.com>