Cygwin: cygthread: Add description of existing cygthreads
We're using the cygthread class in a considerable number of
scenarios, but we don't keep track for which purposes we
use them.

Add a comment at the head of cygthread.cc to describe what
cygthreads we maintain.

This is supposed to be kept up-to-date if some cygthread is
added, modified, or removed (wishful thinking).

Signed-off-by: Corinna Vinschen <corinna@vinschen.de>