Reduce per-target meson code
Make the default directory name '' instead of '.' so that
we can use join_paths(value[0], 'name') everywhere.

Prefix non-default target names with '_' so we can use 'name' + target
everywhere.

Add another value to the target value array which is set to 'lib' for
non-default targets and '' otherwise. With this, we can use
join_paths(value[0], value[2] + 'name') for libraries and ensure that
the basename gets prefixed with 'lib' as needed.

Signed-off-by: Keith Packard <keithp@keithp.com>