[ARM] On Thumb-1 targets, don't overwrite current sp in _set_stacks
In the startup code, when setting the various SPs of the different modes, we
also overwrite the SP of the current mode, which on reset/startup is the
supervisor mode.

Besides this, _set_stacks also switches from supervisor mode to system
mode. Which on first sight seems unintentional, also because this isn't done for
the other targets.

In most cases this happens to work out, because this early in the startup code,
nothing has been written to the stack yet. However when frame pointers for leaf
functions are enabled, which for Clang is now the default when frame pointers
are enabled, we do have stack movement. Both _cstart and _set_stacks copy r11
and lr to the stack on entry and pop them off on exit. And so will write garbage
in the link register.

This patch changes the behaviour of _set_stacks to switches back to supervisor
mode instead of switching to system mode, without overwriting the supervisor
mode SP.