Rework _default_types.h to improve legacy compiler compatibility
This patch improves compatibility with legacy versions of GCC by
deducing the correct value of missing macros.

The __SIZEOF_*__ macros are required by various parts of tinystdio
and was introduced in GCC 4.3.

The __INT{8,16,32,64}__ macros are required by _intsup.h and were
and was introduced in GCC 4.5.

The revised design works as follows:

 1. From __EXP({SCHAR,SHRT,INT}_MAX) and
    __have_{long32,long64,longlong64} determine
    __{SCHAR,SHRT,INT,LONG,LONG_LONG}_WIDTH__ if they are not already
    defined.

 2. From __{SHRT,INT,LONG,LONG_LONG}_WIDTH__ and
    __{FLT,DBL,LDBML}_MANT_DIG__ define
    __SIZEOF_{SHORT,INT,LONG,LONG_LONG,FLOAT,DOUBLE,LONG_DOUBLE}__ if
    they are not already defined.

 3. From __{SHRT,INT,LONG,LONG_LONG}_WIDTH__ define
    __{,U}INT{8,16,32,64}__ if they are not already defined.

Signed-off-by: Joel Holdsworth <jholdsworth@nvidia.com>