Revert "Cygwin: get_posix_access: do not merge permissions for just created files"
This reverts commit 971d2dffea7848270aa9dfb5c14dcd946c8971c0.

This patch was supposed to fix lots of FAILs in our own
testsuite/winsup.api/ltp/umask03.c test.  The reason was
that umask masking in get_posix_access when generating new files
was overriden by later code in the same function, merging
permissions in old-style or Windows generated ACLs.

However, the solution to skip merging was not the right thing,
because this breaks handling of Windows-generated ACLs.

Rather, umask masking should be performed pretty late, certainly
after merging permissions.  This will be done by a followup patch.

Fixes: 971d2dffea78 ("Cygwin: get_posix_access: do not merge permissions for just created files")
Signed-off-by: Corinna Vinschen <corinna@vinschen.de>