Cygwin: get_posix_access: improve merging permissions
Add a second loop to the code snippet merging permissions in old-style
or Windows-generated ACLs.  This loop fixes up default ACL permissions
created from ACEs which are valid for the directory itself, as well as
getting inherited to child objects.

The FULL_ACE bit utilized for this is removed from the ACE at the
end of the function, together with the temporary DENY bits.

Signed-off-by: Corinna Vinschen <corinna@vinschen.de>