Cygwin: get_posix_access: make sure pos is correctly set when used
The code merging permissions relies on `pos' being set to the number
of current entries in the local aclent_t buffer.  Commit 0e6d36766c83
("Cygwin: get_posix_access: move umask masking to the end") moved that
code to run earlier, but neglected to move setting `pos' correctly
as well.

Make sure to set `pos' inside the code block, as well as in the
final array size check, so `pos' is set correctly where it belongs.

Fixes: 0e6d36766c83 ("Cygwin: get_posix_access: move umask masking to the end")
Signed-off-by: Corinna Vinschen <corinna@vinschen.de>