ctype: Limit "C" locale to ASCII in isw* functions
The C language requires that iswblank only return true for ' ' and
'\t' for the C locale. Glibc does by restricting all of these functions
for this locale, so we'll do the same.

This requires passing the locale to __ctype_table_lookup and then
checking the LC_CTYPE locale name to see if it is "C".

Signed-off-by: Keith Packard <keithp@keithp.com>