[riscv/math] Make sqrt function aliases for correct types
Make sqrt function aliases for correct types if double and long double are the
same size, then alias double function to the long double name.

Before this change: building picolibc on RISCV with -Zd -Zf with forcing
math-errno that calls sqrtl() function; the application will fail compilation
as sqrtl is an undefined symbol in this case.

Signed-off-by: Ahmed Shehab <shehab@synopsys.com>