tinylocale: Create minimal locale implementation
This should provide most of the same functionality as the original
locale code as available in embedded systems where there aren't any
locale data files available. It is missing is the ability to construct
an LC_ALL locale string which specifies multiple locale categories,
but as the only useful category is LC_CTYPE, that doesn't seem
terrible useful (and is not required by C).

Signed-off-by: Keith Packard <keithp@keithp.com>