locale: Switch locale from a struct to an int
As the only thing locale controls is LC_CTYPE, get rid of the struct
and replace it with just the LC_CTYPE value. Represent the locale as
an 'int' in the API to avoid ABI compatibility issues with enums.

Because (locale_t) 0 is a special return value from newlocale, shift
all locale ids up by one, leaving 0 for the error return. Use that for
locale_INVALID.

Signed-off-by: Keith Packard <keithp@keithp.com>