ctype: Always allow negative _ctype_ indices
Compilers now provide build-time options controlling whether 'char'
means 'signed char' or 'unsigned char'. This means we cannot assume
the same mode will be used when building picolibc as when building the
application.

To avoid having the ABI shift with compiler flags, remove support for
the smaller tables and always provide tables which allow for signed
char indexing as this provides the safest possible implementation.

Signed-off-by: Keith Packard <keithp@keithp.com>